[REFACTOR] add pushAlarmHandlers
