+++ functions/api/routes/classroomPost/classroomPostPOST.js
@@ -6,8 +6,8 @@ const db = require("../../../db/db");
 const { classroomPostDB, majorDB, userDB, notificationDB } = require("../../../db");
 const notificationType = require("../../../constants/notificationType");
 const postType = require("../../../constants/postType");
-const admin = require("firebase-admin");
 const slackAPI = require("../../../middlewares/slackAPI");
+const pushAlarmHandlers = require("../../../lib/pushAlarmHandlers");
 
 module.exports = async (req, res) => {
   const { majorId, answererId, postTypeId, title, content } = req.body;
@@ -95,39 +95,8 @@ module.exports = async (req, res) => {
           post.postTypeId,
         );
 
-        // 디바이스로 보낼 푸시 알림 메시지
-
-        // 메세지 내용
-        const message = {
-          notification: {
-            title: notificationTitle,
-            body: notificationContent,
-          },
-          android: {
-            notification: {
-              sound: "default",
-            },
-          },
-          apns: {
-            payload: {
-              aps: {
-                sound: "default",
-              },
-            },
-          },
-          token: receiver.deviceToken,
-        };
-
-        // 메세지 전송
-        admin
-          .messaging()
-          .send(message)
-          .then(function (response) {
-            console.log(responseMessage.PUSH_ALARM_SEND_SUCCESS, response);
-          })
-          .catch(function (error) {
-            console.log(responseMessage.PUSH_ALARM_SEND_FAIL, error);
-          });
+        // 푸시 알림 전송
+        pushAlarmHandlers.sendUnicast(receiver.deviceToken, notificationTitle, notificationContent);
       }
     }
