+++ functions/api/routes/comment/commentPOST.js
@@ -6,8 +6,8 @@ const db = require("../../../db/db");
 const { commentDB, userDB, majorDB, classroomPostDB, notificationDB } = require("../../../db");
 const notificationType = require("../../../constants/notificationType");
 const postType = require("../../../constants/postType");
-const admin = require("firebase-admin");
 const slackAPI = require("../../../middlewares/slackAPI");
+const pushAlarmHandlers = require("../../../lib/pushAlarmHandlers");
 
 module.exports = async (req, res) => {
   const { postId, content } = req.body;
@@ -100,39 +100,8 @@ module.exports = async (req, res) => {
           commentPost.postTypeId,
         );
 
-        // 디바이스로 보낼 푸시 알림 메시지
-
-        // 메세지 내용
-        const message = {
-          notification: {
-            title: notificationTitle,
-            body: notificationContent,
-          },
-          android: {
-            notification: {
-              sound: "default",
-            },
-          },
-          apns: {
-            payload: {
-              aps: {
-                sound: "default",
-              },
-            },
-          },
-          token: receiver.deviceToken,
-        };
-
-        // 메세지 전송
-        admin
-          .messaging()
-          .send(message)
-          .then(function (response) {
-            console.log(responseMessage.PUSH_ALARM_SEND_SUCCESS, response);
-          })
-          .catch(function (error) {
-            console.log(responseMessage.PUSH_ALARM_SEND_FAIL, error);
-          });
+        // 푸시 알림 전송
+        pushAlarmHandlers.sendUnicast(receiver.deviceToken, notificationTitle, notificationContent);
       }
     }
 
@@ -157,39 +126,8 @@ module.exports = async (req, res) => {
           commentPost.postTypeId,
         );
 
-        // 디바이스로 보낼 푸시 알림 메시지
-
-        // 메세지 내용
-        const message = {
-          notification: {
-            title: notificationTitle,
-            body: notificationContent,
-          },
-          android: {
-            notification: {
-              sound: "default",
-            },
-          },
-          apns: {
-            payload: {
-              aps: {
-                sound: "default",
-              },
-            },
-          },
-          token: receiver.deviceToken,
-        };
-
-        // 메세지 전송
-        admin
-          .messaging()
-          .send(message)
-          .then(function (response) {
-            console.log(responseMessage.PUSH_ALARM_SEND_SUCCESS, response);
-          })
-          .catch(function (error) {
-            console.log(responseMessage.PUSH_ALARM_SEND_FAIL, error);
-          });
+        // 푸시 알림 전송
+        pushAlarmHandlers.sendUnicast(receiver.deviceToken, notificationTitle, notificationContent);
       }
     }
 
@@ -233,61 +171,14 @@ module.exports = async (req, res) => {
               comment.commentId,
               commentPost.postTypeId,
             );
+            // 댓글 리스트에 있는 유저들의 디바이스 토큰 정보 저장
             receiverTokens.push(receiver.deviceToken);
           }
         }),
       );
 
-      // 디바이스로 보낼 푸시 알림 메시지
-
-      // 댓글 리스트에 있는 유저들의 디바이스 토큰 정보 저장
-      // 댓글이 있을 때만 푸시알림 전송, 댓글이 없을 경우 tokens가 빈 배열이라서 오류남.
-
-      if (receiverTokens.length !== 0) {
-        // 메세지 내용
-        const message = {
-          notification: {
-            title: notificationTitle,
-            body: notificationContent,
-          },
-          android: {
-            notification: {
-              sound: "default",
-            },
-          },
-          apns: {
-            payload: {
-              aps: {
-                sound: "default",
-              },
-            },
-          },
-          tokens: receiverTokens,
-        };
-        // 메세지 전송
-        admin
-          .messaging()
-          .sendMulticast(message)
-          .then((response) => {
-            console.log(responseMessage.PUSH_ALARM_SEND_SUCCESS, response.successCount);
-            if (response.failureCount > 0) {
-              const failedTokens = [];
-              response.responses.forEach((response, i) => {
-                if (!response.success) {
-                  failedTokens.push(receiverTokens[i]);
-                }
-              });
-              console.log(
-                responseMessage.PUSH_ALARM_SEND_FAIL,
-                response.failureCount,
-                failedTokens,
-              );
-            }
-          })
-          .catch(function (error) {
-            console.log(responseMessage.PUSH_ALARM_SEND_FAIL, error);
-          });
-      }
+      // 푸시 알림 전송
+      pushAlarmHandlers.sendMulticast(receiverTokens, notificationTitle, notificationContent);
     }
 
     // [ case 5: 내가 답글을 쓴 타인 글에 새 답글이 달린 경우 - 정보글 ]
@@ -327,62 +218,14 @@ module.exports = async (req, res) => {
               comment.commentId,
               commentPost.postTypeId,
             );
+            // 댓글 리스트에 있는 유저들의 디바이스 토큰 정보 저장
             receiverTokens.push(receiver.deviceToken);
           }
         }),
       );
 
-      // 디바이스로 보낼 푸시 알림 메시지
-
-      // 댓글 리스트에 있는 유저들의 디바이스 토큰 정보 저장
-      // 댓글이 있을 때만 푸시알림 전송, 댓글이 없을 경우 tokens가 빈 배열이라서 오류남.
-
-      if (receiverTokens.length !== 0) {
-        // 메세지 내용
-        const message = {
-          notification: {
-            title: notificationTitle,
-            body: notificationContent,
-          },
-          android: {
-            notification: {
-              sound: "default",
-            },
-          },
-          apns: {
-            payload: {
-              aps: {
-                sound: "default",
-              },
-            },
-          },
-          tokens: receiverTokens,
-        };
-
-        // 메세지 전송
-        admin
-          .messaging()
-          .sendMulticast(message)
-          .then((response) => {
-            console.log(responseMessage.PUSH_ALARM_SEND_SUCCESS, response.successCount);
-            if (response.failureCount > 0) {
-              const failedTokens = [];
-              response.responses.forEach((response, i) => {
-                if (!response.success) {
-                  failedTokens.push(receiverTokens[i]);
-                }
-              });
-              console.log(
-                responseMessage.PUSH_ALARM_SEND_FAIL,
-                response.failureCount,
-                failedTokens,
-              );
-            }
-          })
-          .catch(function (error) {
-            console.log(responseMessage.PUSH_ALARM_SEND_FAIL, error);
-          });
-      }
+      // 푸시 알림 전송
+      pushAlarmHandlers.sendMulticast(receiverTokens, notificationTitle, notificationContent);
     }
 
     res
